eduflow_sso_web_app/pyproject.toml
[project]
name = "codeline-sso-web-app"
version = "0.1.0"
description = "FastAPI Microsoft SSO (OAuth 2.0 / OIDC) application"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.34.2",
    "authlib>=1.5.2",
    "joserfc>=0.12.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "python-dotenv>=1.1.0",
    "aiofiles>=24.1.0",
    "ffmpeg-python>=0.2.0",
    "aiosqlite>=0.20.0",
    "elevenlabs>=2.0.0",
    "boto3>=1.35.0",
    "psycopg2-binary>=2.9.0",
]

[dependency-groups]
dev = []

[tool.setuptools.packages.find]
include = ["app*"]

[tool.uv]
package = true

[project.scripts]
serve = "app.main:run"
dev = "app.main:dev"
